Force to reload the apparmor profile if profile(docker-default) is installed previously.
docker-snap has a chance to load the old apparmor profile(docker-default) if
docker deb package is installed previously, that will cause IPC failed to
execute via signals.
This commit enables docker forced to reload the apparmor profile if a
profile(docker-default) exists before starting docker daemon. And the apparmor
profile with the same name will be refreshed if it's the case and the
right profile for docker snap will be loaded correctly then.)
Fixed #4