Add a patch to make certificates directory accessable in snappy world.
For the sake of easy setup for secure registry server, the manual steps
for signed-certificate generation are not required, snap is in charge of
certificate generation right after the first installation.
There're two available configuration options can be set-able via a config hook.
One is registry-domain-name, another is certificate-folder. Please take
a look at docker.help for further details.