Build golang1.8.3 from source.
As a build pre-dependencies, starting from docker 17.06, golang 1.8.3
is adopted as the default compiler to build docker. In the short term,
we build the golang 1.8.3 from source and in the long run, we need to
get golang 1.8.3 back-ported from ubuntu achieve so that we can benefit
from the security patches.