Use aufs as default strage driver.
From 17.09, overlay2 is being used as default storage driver instead of aufs.
see this PR for more details
https://github.com/docker/docker.github.io/issues/4141
However when running the spread test on ubuntu-core, we see the apparmor denials
(the root path contains the rootfs partition"system-data") after switching to overlay2.
So for this release, we still stick around aufs as the default storage driver.